React and Redux: Manage State Locally First
First, in the render()method, have the component render an inputelement, buttonelement, and ulelement. When the inputelement changes, it should trigger a handleChange()method. Also, the inputelement should render the value of inputthat's in the component's state. The buttonelement should trigger a submitMessage()method when it's clicked.

Second, write these two methods. The handleChange()method should update the inputwith what the user is typing. The submitMessage()method should concatenate the current message (stored in input) to the messagesarray in local state, and clear the value of the input.