React and Redux: Connect Redux to the Messages App
The code editor has all the code you've written in this section so far. The only change is that the React component is renamed to Presentational. Create a new component held in a constant called Containerthat uses connectto connect the Presentationalcomponent to Redux. Then, in the AppWrapper, render the React Redux Providercomponent. Pass Providerthe Redux storeas a prop and render Containeras a child. Once everything is setup, you will see the messages app rendered to the page again.